requirements.txt
numpy>=1.18.0
pygame>=2.0.0
PyOpenGL>=3.1.0
screeninfo>=0.6.0
sounddevice
